+++ noxfile.py
@@ -2,12 +2,12 @@
 
 #  ************************** Copyrights and license ***************************
 #
-# This file is part of gcovr 5.2, a parsing and reporting tool for gcov.
+# This file is part of gcovr 6.0+master, a parsing and reporting tool for gcov.
 # https://gcovr.com/en/stable
 #
 # _____________________________________________________________________________
 #
-# Copyright (c) 2013-2022 the gcovr authors
+# Copyright (c) 2013-2023 the gcovr authors
 # Copyright (c) 2013 Sandia Corporation.
 # Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
 # the U.S. Government retains certain rights in this software.
@@ -58,6 +58,16 @@ DEFAULT_LINT_ARGUMENTS = [
 ] + DEFAULT_TEST_DIRECTORIES
 
 BLACK_PINNED_VERSION = "black==22.3.0"
+
+OUTPUT_FORMATS = [
+    "cobertura",
+    "coveralls",
+    "csv",
+    "html-details",
+    "json",
+    "sonarqube",
+    "txt",
+]
 
 nox.options.sessions = ["qa"]
 
@@ -113,6 +123,13 @@ def set_environment(session: nox.Session, cc: str, check: bool = True) -> None:
     session.env["CFLAGS"] = "--this_flag_does_not_exist"
     session.env["CXX"] = cxx
     session.env["CXXFLAGS"] = "--this_flag_does_not_exist"
+
+
+@nox.session
+def bump_version(session: nox.Session) -> None:
+    """Bump the new version"""
+    session.install("-e", ".")
+    session.run("python", "admin/bump_version.py")
 
 
 @nox.session(python=False)
@@ -243,9 +260,6 @@ def tests_compiler(session: nox.Session, version: str) -> None:
     args = ["-m", "pytest"]
     args += coverage_args
     args += session.posargs
-    # For docker tests
-    if "NOX_POSARGS" in os.environ:
-        args += shlex.split(os.environ["NOX_POSARGS"])
     if "--" not in args:
         args += ["--"] + DEFAULT_TEST_DIRECTORIES
     session.run("python", *args)
@@ -266,11 +280,15 @@ def build_wheel(session: nox.Session) -> None:
 def check_wheel(session: nox.Session) -> None:
     """Check the wheel and do a smoke test, should not be used directly."""
     session.install("wheel", "twine")
-    session.chdir(f"{session.cache_dir}/dist")
-    session.run("twine", "check", "*", external=True)
-    session.install(glob.glob("*.whl")[0])
+    with session.chdir(f"{session.cache_dir}/dist"):
+        session.run("twine", "check", "*", external=True)
+        session.install(glob.glob("*.whl")[0])
     session.run("python", "-m", "gcovr", "--help", external=True)
     session.run("gcovr", "--help", external=True)
+    session.log("Run all transformations to check if all the modules are packed")
+    with session.chdir(session.create_tmp()):
+        for format in OUTPUT_FORMATS:
+            session.run("gcovr", f"--{format}", f"out.{format}", external=True)
 
 
 @nox.session
@@ -301,7 +319,7 @@ def bundle_app(session: nox.Session) -> None:
         "./pyinstaller",
         "--onefile",
         "--collect-all",
-        "gcovr",
+        "gcovr.writer",
         "-n",
         executable,
         *session.posargs,
@@ -313,11 +331,18 @@ def bundle_app(session: nox.Session) -> None:
 @nox.session
 def check_bundled_app(session: nox.Session) -> None:
     """Run a smoke test with the bundled app, should not be used directly."""
-    session.chdir("build")
-    session.run("bash", "-c", "./gcovr --help", external=True)
-    session.log("Run HTML all transformations to check if all the modules are packed")
-    for format in ["txt", "html", "cobertura", "sonarqube", "csv", "coveralls"]:
-        session.run("bash", "-c", f"./gcovr --{format} out.{format}", external=True)
+    with session.chdir("build"):
+        # bash here is needed to be independent from the file extension (Windows).
+        session.run("bash", "-c", "./gcovr --help", external=True)
+        session.log("Run all transformations to check if all the modules are packed")
+        session.create_tmp()
+        for format in OUTPUT_FORMATS:
+            session.run(
+                "bash",
+                "-c",
+                f"./gcovr --{format} $TMPDIR/out.{format}",
+                external=True,
+            )
 
 
 def docker_container_os(session: nox.Session) -> str:
@@ -334,25 +359,25 @@ def docker_container_id(session: nox.Session, version: str) -> str:
 
 
 @nox.session(python=False)
-def docker_qa_build(session: nox.Session) -> None:
+def docker_build(session: nox.Session) -> None:
     """Build the docker container for the default GCC version."""
-    session_id = f"docker_qa_build({GCC_VERSIONS[0]})"
+    session_id = f"docker_build({GCC_VERSIONS[0]})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
 
 
-@nox.session(python=False, name="docker_qa_build_compiler(all)")
-def docker_qa_build_compiler_all(session: nox.Session) -> None:
+@nox.session(python=False, name="docker_build_compiler(all)")
+def docker_build_compiler_all(session: nox.Session) -> None:
     """Build the docker containers vor all GCC versions."""
     for version in GCC_VERSIONS:
-        session_id = f"docker_qa_build_compiler({version})"
+        session_id = f"docker_build_compiler({version})"
         session.log(f"Notify session {session_id}")
         session.notify(session_id)
 
 
 @nox.session(python=False)
 @nox.parametrize("version", [nox.param(v, id=v) for v in GCC_VERSIONS])
-def docker_qa_build_compiler(session: nox.Session, version: str) -> None:
+def docker_build_compiler(session: nox.Session, version: str) -> None:
     """Build the docker container for a specific GCC version."""
     set_environment(session, version, False)
     session.run(
@@ -376,25 +401,25 @@ def docker_qa_build_compiler(session: nox.Session, version: str) -> None:
 
 
 @nox.session(python=False)
-def docker_qa_run(session: nox.Session) -> None:
+def docker_run(session: nox.Session) -> None:
     """Run the docker container for the default GCC version."""
-    session_id = f"docker_qa_run_compiler({GCC_VERSIONS[0]})"
+    session_id = f"docker_run_compiler({GCC_VERSIONS[0]})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
 
 
-@nox.session(python=False, name="docker_qa_run_compiler(all)")
-def docker_qa_run_compiler_all(session: nox.Session) -> None:
+@nox.session(python=False, name="docker_run_compiler(all)")
+def docker_run_compiler_all(session: nox.Session) -> None:
     """Run the docker container for the all GCC versions."""
     for version in GCC_VERSIONS:
-        session_id = f"docker_qa_run_compiler({version})"
+        session_id = f"docker_run_compiler({version})"
         session.log(f"Notify session {session_id}")
         session.notify(session_id)
 
 
 @nox.session(python=False)
 @nox.parametrize("version", [nox.param(v, id=v) for v in GCC_VERSIONS])
-def docker_qa_run_compiler(session: nox.Session, version: str) -> None:
+def docker_run_compiler(session: nox.Session, version: str) -> None:
     """Run the docker container for a specific GCC version."""
     set_environment(session, version, False)
 
@@ -405,13 +430,11 @@ def docker_qa_run_compiler(session: nox.Session, version: str) -> None:
             # Code for join taken from Python 3.9
             return " ".join(shlex.quote(arg) for arg in args)
 
-    session.env["NOX_POSARGS"] = shell_join(session.posargs)
-    nox_options = []
+    nox_options = session.posargs
     if session._runner.global_config.no_install:
-        nox_options.append("--no-install")
+        nox_options.insert(0, "--no-install")
     if session._runner.global_config.reuse_existing_virtualenvs:
-        nox_options.append("--reuse-existing-virtualenvs")
-    session.env["NOX_OPTIONS"] = shell_join(nox_options)
+        nox_options.insert(0, "--reuse-existing-virtualenvs")
     session.run(
         "docker",
         "run",
@@ -419,20 +442,46 @@ def docker_qa_run_compiler(session: nox.Session, version: str) -> None:
         "-t",
         "-e",
         "CC",
-        "-e",
-        "NOX_POSARGS",
-        "-e",
-        "NOX_OPTIONS",
         "-v",
         f"{os.getcwd()}:/gcovr",
         docker_container_id(session, version),
+        *nox_options,
         external=True,
     )
 
 
 @nox.session(python=False)
-def docker_qa(session: nox.Session) -> None:
+def docker(session: nox.Session) -> None:
     """Build and run the docker container for the default GCC version."""
+    session_id = f"docker_compiler({GCC_VERSIONS[0]})"
+    session.log(f"Notify session {session_id}")
+    session.notify(session_id)
+
+
+@nox.session(python=False, name="docker_compiler(all)")
+def docker_compiler_all(session: nox.Session) -> None:
+    """Build and run the docker container for all GCC versions."""
+    for version in GCC_VERSIONS:
+        session_id = f"docker_compiler({version})"
+        session.log(f"Notify session {session_id}")
+        session.notify(session_id)
+
+
+@nox.session(python=False)
+@nox.parametrize("version", [nox.param(v, id=v) for v in GCC_VERSIONS])
+def docker_compiler(session: nox.Session, version: str) -> None:
+    """Build and run the docker container for a specific GCC version."""
+    session_id = "docker_build_compiler({})".format(version)
+    session.log(f"Notify session {session_id}")
+    session.notify(session_id)
+    session_id = f"docker_run_compiler({version})"
+    session.log(f"Notify session {session_id}")
+    session.notify(session_id, session.posargs)
+
+
+@nox.session(python=False)
+def docker_qa(session: nox.Session) -> None:
+    """Run the session qa for the default GCC version."""
     session_id = f"docker_qa_compiler({GCC_VERSIONS[0]})"
     session.log(f"Notify session {session_id}")
     session.notify(session_id)
@@ -440,7 +489,7 @@ def docker_qa(session: nox.Session) -> None:
 
 @nox.session(python=False, name="docker_qa_compiler(all)")
 def docker_qa_compiler_all(session: nox.Session) -> None:
-    """Build and run the docker container for the all GCC versions."""
+    """Run the session qa for all GCC versions."""
     for version in GCC_VERSIONS:
         session_id = f"docker_qa_compiler({version})"
         session.log(f"Notify session {session_id}")
@@ -450,10 +499,7 @@ def docker_qa_compiler_all(session: nox.Session) -> None:
 @nox.session(python=False)
 @nox.parametrize("version", [nox.param(v, id=v) for v in GCC_VERSIONS])
 def docker_qa_compiler(session: nox.Session, version: str) -> None:
-    """Build and run the docker container for a specific GCC version."""
-    session_id = "docker_qa_build_compiler({})".format(version)
+    """Run the session qa for a specific GCC version."""
+    session_id = f"docker_compiler({version})"
     session.log(f"Notify session {session_id}")
-    session.notify(session_id)
-    session_id = f"docker_qa_run_compiler({version})"
-    session.log(f"Notify session {session_id}")
-    session.notify(session_id)
+    session.notify(session_id, ["-s", "qa", "--", *session.posargs])